added 2 new components and 2 new css files
Added the gamesDetails.jsx component and the Games.jsx component. The games component shows all the games and it includes a search bar. the gamesdetails page is used when the user clicks on a game to review more info. I added two css files to coordinate along with these components.